Use is_persistent=1 for superglobal name's zend_string, like rest of superglobals
See the other calls to zend_register_auto_global, they all pass in
persistent strings. Add a test of the expected behaviour, with various
ways to access these persistent strings.

NTS and ZTS are different - NTS does not have a interned strings.

The way that interned strings are created differs in NTS and ZTS builds,
so use %d for EXPECTF in the test.

Update appveyor to install $_SESSION, since it's not possible to disable
creating $_SESSION in php.ini.