Fix a crash in runkit_function_rename
Fixes #153

The stack size could end up incorrect if runkit_function_rename was used
to replace the original removed/renamed function definition
with a different renamed function definition.

Unlike runkit_function_add and runkit_function_redefine,
the stack size fix wasn't getting applied for runkit_function_rename.

This PR fixes that by applying the fix to increase the maximum stack
size to be large enough for the **new** renamed definition of the
function, for anything that was referring to a different global function
definition with a potentially smaller stack size.